+++ tests/test_modules.py
@@ -60,7 +60,7 @@ class TestModules(unittest.TestCase):
         assert modrepo.name == "nf-core/modules"
         assert modrepo.branch == "master"
 
-    def test_modules_list(self):
+    def test_modules_list_remote(self):
         """Test listing available modules"""
         mods_list = nf_core.modules.ModuleList(None)
         listed_mods = mods_list.list_modules()
@@ -68,6 +68,26 @@ class TestModules(unittest.TestCase):
         console.print(listed_mods)
         output = console.export_text()
         assert "fastqc" in output
+
+    def test_modules_list_pipeline(self):
+        """Test listing locally installed modules"""
+        mods_list = nf_core.modules.ModuleList(self.pipeline_dir)
+        listed_mods = mods_list.list_modules()
+        console = Console(record=True)
+        console.print(listed_mods)
+        output = console.export_text()
+        assert "fastqc" in output
+        assert "multiqc" in output
+
+    def test_modules_install_and_list_pipeline(self):
+        """Test listing locally installed modules"""
+        self.mods_install.install("trimgalore")
+        mods_list = nf_core.modules.ModuleList(self.pipeline_dir)
+        listed_mods = mods_list.list_modules()
+        console = Console(record=True)
+        console.print(listed_mods)
+        output = console.export_text()
+        assert "trimgalore" in output
 
     def test_modules_install_nopipeline(self):
         """Test installing a module - no pipeline given"""